+++ LendingWebApp/Controllers/HomeController.cs
@@ -1,8 +1,7 @@
 using System.Diagnostics;
-using Loan_application_service.Models;
 using Microsoft.AspNetCore.Mvc;
 
-namespace Loan_application_service.Controllers
+namespace LoanApplicationService.Web.Controllers
 {
     public class HomeController : Controller
     {
@@ -23,10 +22,10 @@ namespace Loan_application_service.Controllers
             return View();
         }
 
-        [ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
-        public IActionResult Error()
-        {
-            return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
-        }
+        //[ResponseCache(Duration = 0, Location = ResponseCacheLocation.None, NoStore = true)]
+        //public IActionResult Error()
+        //{
+        //    return View(new ErrorViewModel { RequestId = Activity.Current?.Id ?? HttpContext.TraceIdentifier });
+        //}
     }
 }
