--- LendingWebApp/DTOs/UserDTO.cs
@@ -1,13 +1,0 @@
-namespace Loan_application_service.Models;
-
-public class UserDTO
-{
-    public int UserId { get; set; }
-    public string? Username { get; set; }
-    public string? Email { get; set; }
-    public string? PasswordHash { get; set; }
-    public string? Role { get; set; }
-    public string? FirstName { get; set; }
-    public string? LastName { get; set; }
-    public bool IsActive { get; set; }
-}
