+++ LendingWebApp/LoanApplicationService.Web.csproj
@@ -1,0 +1,18 @@
+<Project Sdk="Microsoft.NET.Sdk.Web">
+
+  <PropertyGroup>
+    <TargetFramework>net9.0</TargetFramework>
+    <Nullable>enable</Nullable>
+    <ImplicitUsings>enable</ImplicitUsings>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="9.0.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\LoanApplicationService.Core\LoanApplicationService.Core.csproj" />
+    <ProjectReference Include="..\LoanApplicationService.Service\LoanApplicationService.Service.csproj" />
+  </ItemGroup>
+
+</Project>
