--- LendingWebApp/Models/ErrorViewModel.cs
@@ -1,9 +1,0 @@
-namespace Loan_application_service.Models
-{
-    public class ErrorViewModel
-    {
-        public string? RequestId { get; set; }
-
-        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
-    }
-}
