--- LendingWebApp/Models/User.cs
@@ -1,47 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-
-namespace Loan_application_service.Models
-{
-    public class Users
-    {
-        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        [Key]
-        public int UserId { get; set; }
-        
-        [Required]
-        [MaxLength(50)]
-        public string? Username { get; set; }
-        
-        [Required]
-        [MaxLength(100)]
-        [EmailAddress]
-        public string? Email { get; set; }
-        
-        [Required]
-        [MaxLength(255)]
-        public string? PasswordHash { get; set; }
-        
-        [Required]
-        [MaxLength(20)]
-        public string? Role { get; set; }
-        
-        [MaxLength(50)]
-        public string? FirstName { get; set; }
-        
-        [MaxLength(50)]
-        public string? LastName { get; set; }
-        
-        public bool IsActive { get; set; } = true;
-        
-        public DateTime? LastLogin { get; set; }
-        
-        public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
-        public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
-        
-        // Navigation Properties
-        public virtual ICollection<LoanApplication> ProcessedApplications { get; set; } = new List<LoanApplication>();
-        public virtual ICollection<Notification> SentNotifications { get; set; } = new List<Notification>();
-        public virtual ICollection<AuditTrail> AuditTrails { get; set; } = new List<AuditTrail>();
-    }
-}
