+++ LendingWebApp/Views/Home/Index.cshtml
@@ -1,0 +1,1 @@
+﻿<h1>Loans Lending System</h1>
