+++ LendingWebApp/Views/LoanProduct/Index.cshtml
@@ -1,0 +1,49 @@
+﻿@model IEnumerable<LoanApplicationService.Service.DTOs.LoanModule.LoanProductDto>
+@{
+    ViewBag.Title = "All Products";
+}
+
+<h2>Loan Products</h2>
+
+<div class="mb-3">
+    <a href="@Url.Action("Create", "LoanProduct")" class="btn btn-primary">Create New Product</a>
+</div>
+
+<table class="table table-bordered table-striped">
+    <thead class="table-dark">
+        <tr>
+            <th>Product Name</th>
+            <th>Product Type</th>
+            <th>Min Amount</th>
+            <th>Max Amount</th>
+            <th>Interest Rate (%)</th>
+            <th>Min Term (Months)</th>
+            <th>Max Term (Months)</th>
+            <th>Processing Fee</th>
+            <th>Eligibility</th>
+            <th>Actions</th>
+        </tr>
+    </thead>
+    <tbody>
+        @foreach (var product in Model)
+        {
+            <tr>
+                <td>@product.ProductName</td>
+                <td>@product.ProductType</td>
+                <td>@product.MinAmount.ToString("N2")</td>
+                <td>@product.MaxAmount.ToString("N2")</td>
+                <td>@(product.InterestRate * 100).ToString("F2")</td>
+                <td>@product.MinTermMonths</td>
+                <td>@product.MaxTermMonths</td>
+                <td>@product.ProcessingFee.ToString("N2")</td>
+                <td>@product.EligibilityCriteria</td>
+                <td>
+                    <a class="btn btn-info btn-sm" href="@Url.Action("Details", "LoanProduct", new { id = product.ProductId })">View</a>
+                    <a class="btn btn-warning btn-sm" href="@Url.Action("Edit", "LoanProduct", new { id = product.ProductId })">Edit</a>
+                </td>
+            </tr>
+        }
+    </tbody>
+</table>
+
+
