+++ LendingWebApp/appsettings.json
@@ -7,6 +7,6 @@
   },
   "AllowedHosts": "*",
   "ConnectionStrings": {
-    "Loan_application_serviceContext": "Server=DESKTOP-GCP876V\\SQLEXPRESS;Database=Loan_application_service;Trusted_Connection=True;MultipleActiveResultSets=true;TrustServerCertificate=true"
+    "DbConnection": "Server=DESKTOP-GCP876V\\SQLEXPRESS;Database=Loan_application_service;Trusted_Connection=True;MultipleActiveResultSets=true;TrustServerCertificate=true"
   }
 }
