+++ LoanApplicationService.Core/Models/ErrorViewModel.cs
@@ -1,0 +1,9 @@
+namespace LoanApplicationService.Core.Models
+{
+    public class ErrorViewModel
+    {
+        public string? RequestId { get; set; }
+
+        public bool ShowRequestId => !string.IsNullOrEmpty(RequestId);
+    }
+}
