+++ LoanApplicationService.Core/Models/User.cs
@@ -1,0 +1,47 @@
+﻿using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace LoanApplicationService.Core.Models
+{
+    public class Users
+    {
+        [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
+        [Key]
+        public int UserId { get; set; }
+
+        [Required]
+        [MaxLength(50)]
+        public string? Username { get; set; }
+
+        [Required]
+        [MaxLength(100)]
+        [EmailAddress]
+        public string? Email { get; set; }
+
+        [Required]
+        [MaxLength(255)]
+        public string? PasswordHash { get; set; }
+
+        [Required]
+        [MaxLength(20)]
+        public string? Role { get; set; }
+
+        [MaxLength(50)]
+        public string? FirstName { get; set; }
+
+        [MaxLength(50)]
+        public string? LastName { get; set; }
+
+        public bool IsActive { get; set; } = true;
+
+        public DateTime? LastLogin { get; set; }
+
+        public DateTime CreatedAt { get; set; } = DateTime.UtcNow;
+        public DateTime UpdatedAt { get; set; } = DateTime.UtcNow;
+
+        // Navigation Properties
+        public virtual ICollection<LoanApplication> ProcessedApplications { get; set; } = [];
+        public virtual ICollection<Notification> SentNotifications { get; set; } = [];
+        public virtual ICollection<AuditTrail> AuditTrails { get; set; } = [];
+    }
+}
