+++ LoanApplicationService.Service/DTOs/UserModule/UserDTO.cs
@@ -1,0 +1,13 @@
+namespace LoanApplicationService.Service.DTOs.UserModule;
+
+public class UserDTO
+{
+    public int UserId { get; set; }
+    public string? Username { get; set; }
+    public string? Email { get; set; }
+    public string? PasswordHash { get; set; }
+    public string? Role { get; set; }
+    public string? FirstName { get; set; }
+    public string? LastName { get; set; }
+    public bool IsActive { get; set; }
+}
