+++ LoanApplicationService.Service/Services/ILoanApplicationService.cs
@@ -1,0 +1,7 @@
+﻿using System;
+
+namespace LoanApplicationService.Service.Services;
+public interface ILoanApplicationService
+{
+
+}
