+++ LoanApplicationService.Service/Services/LoanApplicationServiceImpl.cs
@@ -1,0 +1,6 @@
+﻿using System;
+namespace LoanApplicationService.Service.Services;
+public class LoanApplicationServiceImpl : ILoanApplicationService
+{
+	
+}
